[Fix] Post, Category 로딩 안되는 문제 해결
