+++ src/App.js
@@ -1,5 +1,11 @@
-import { Routes, Route } from 'react-router';
+import { useEffect } from 'react';
+import { Routes, Route, useLocation } from 'react-router';
+import { useDispatch } from 'react-redux';
 
+import { getPostAPI } from '@services/post-api';
+import { initPost } from '@store/post-store';
+import { getCategoriesAPI } from './services/category-api';
+import { initCategories } from './store/category-store';
 import StyledApp from './styles/App-styled';
 import Main from './pages/main/Index';
 import Create from './pages/create/Create';
@@ -7,6 +13,29 @@ import Update from './pages/update/Update';
 import Detail from './pages/main/post/Detail';
 
 const App = () => {
+  const dispatch = useDispatch();
+  const location = useLocation();
+
+  useEffect(() => {
+    const getPost = async () => {
+      const data = await getPostAPI();
+
+      dispatch(initPost(data));
+    };
+
+    if (location.search.length === 0) getPost();
+  }, [location, dispatch]);
+
+  useEffect(() => {
+    const getCategories = async () => {
+      const categoriesData = await getCategoriesAPI();
+
+      dispatch(initCategories(categoriesData));
+    };
+
+    getCategories();
+  }, [dispatch]);
+
   return (
     <StyledApp>
       <Routes>
